Solar panels have not more information in gui

+++ src/main/java/info/u_team/hycrafthds_wtf_ic2_addon/config/ConfigEntrySolarPanel.java
@@ -24,7 +24,7 @@ public class ConfigEntrySolarPanel {
 	@Name("Capacity")
 	public double capacity;
 	
-	public ConfigEntrySolarPanel(int generateDay, int generateNight, int tier, int output, int capacity) {
+	public ConfigEntrySolarPanel(int generateDay, int generateNight, int tier, int output, double capacity) {
 		this.generateDay = generateDay;
 		this.generateNight = generateNight;
 		this.tier = tier;
